Create menu escolar
Simple menu and sub menu with funcionalities choosed by number inputs, with invalid number input treatment. it can add a student, modify the test note, exclude a student, show a complete relatory of all students, show a complete relatory of a single student and finalize the menu and the session. it is not linked with a DB, so, at each use, needs to do all the inputs.